Limiter
 - poprawka limitera

+++ source/mtAudioEngine.h
@@ -26,7 +26,7 @@ public:
 	void setReverbDamping(uint8_t value);
 	void setReverbPanning(int8_t value);
 	void setLimiterAttack(uint16_t attack);
-	void setLimiterRelease(uint16_t release);
+	void setLimiterRelease(float release);
 	void setLimiterTreshold(uint16_t threshold);
 	void muteTrack(uint8_t channel);
 	void wavExportConnect();
